Refactor: hook폴더 안에 서버에서 이미지 데이터 가져오는 로직을 useFetchData.js 훅으로 분리
